Created Retrofit adapter

+++ app/src/main/java/com/testapp/challenge/model/network/dto/Point.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
  * @author aliakseicherniakovich
  */
 class Point(
-    @SerializedName("x") var x: Double,
-    @SerializedName("y") val y: Double
+    @SerializedName("x") var x: Float,
+    @SerializedName("y") val y: Float
 )
